Assignment 1 PHP: Temperature Conversion
On the Fahrenheit temperature scale, water freezes at 32 degrees and boils at 212 degrees. On the Celsius scale, water freezes at 0 degrees and boils at 100 degrees. If you want to convert from Fahrenheit to Celsius, you subtract 32 from the temperature, multiply by 5, and divide by 9. To convert from Celsius to Fahrenheit, you multiply by 9, divide by 5, and then add 32.

Write a PHP program asg1.php that will display a form asking for four parameters:
Convert “F to C” or “C to F” (in a drop-down list – see lab 8 on what is a drop-down list)
the starting Fahrenheit/Celsius degree
the ending Fahrenheit/Celsius degree
the increment in Fahrenheit/Celsius
Then the program asg1.php will call itself. That is, the action part of the form is the same file name. The PHP program file name can be changed and yet you do not need to change the source code. (Look at the source of sample6.php on how a php program calls itself and how to find out its own file name).

Upon the second entry of the program asg1.php, it will print out a table of Fahrenheit and Celsius temperature equivalents from starting degrees to ending degrees with increment degree.
For example, if your drop down list is “F to C” and input starting value = 0; ending value = 100 and increment = 5 the output will be similar to the one shown above. You may use floating point instead of integer and the display can be 0.0 instead of 0. That is the display is similar and not necessarily totally identical.

To format the output nicely, you may need to use <table> <td> <tr> etc.
Do not use spaces or tab to do the format as the browser will treat multiple spaces as a single space.

The file should be “a single file” called asg1.php and submit it to buelearning.
Hints to assignment 1:
The following is a sample PHP program that prints a multiplication table. It is just to show you how to create an HTML table using PHP